+++ tests/test_lines.py
@@ -2,7 +2,7 @@
 __author__ = Hagai Hargil
 """
 from unittest import TestCase
-from pysight.rectify_lines import LineRectifier
+from pysight.nd_hist_generator.rectify_lines import LineRectifier
 import numpy as np